+++ helpers/touchscreen.ts
@@ -1,0 +1,6 @@
+/**
+ * Check if touch screen
+ * @returns {boolean}
+ */
+export const isTouchScreen = (): boolean =>
+  'ontouchstart' in document.documentElement