+++ CHANGELOG.md
@@ -5,6 +5,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.2.0] - 2020-04-24
+
+### Added
+
+- Show player controls on fullscreen
+- Twitter link below player
+
+### Fixed
+
+- Player don't go fullscreen (with duration shown) anymore on iOS
+
 ## [1.1.2] - 2020-04-20
 
 ### Changed